refactor(P1): complete custom hook consolidation - inline useKeyboardNavigationProtection
- Inline useKeyboardNavigationProtection logic in SessionTablePage.tsx and SessionDetail.tsx
- Remove hook file as it was low complexity (15 lines) with limited reuse (2 components)
- Update test files to remove hook mocks
- All format, lint, and type checks pass
- Follows React coding standards: "Create custom hooks only for truly reusable functionality"
