+++ humanlayer-wui/REACT_REFACTOR_PLAN.md
@@ -432,17 +432,19 @@ src/components/
 
 **Status**: Additional comprehensive implementation completed  
 **Time Spent**: ~6 additional hours  
-**Files Changed**: 10+ additional files  
+**Files Changed**: 10+ additional files
 
 #### What Was Added Beyond Previous Implementation:
 
 1. **Created Comprehensive Test Infrastructure**:
+
    - ✅ **UI Slice Tests** (`src/stores/uiSlice.test.ts`): 13 tests covering modal state, editing state, and UI integration
    - ✅ **SessionTable Integration Tests** (`src/pages/SessionTablePage.test.tsx`): 10 tests for navigation, selection, view modes, and operations
    - ✅ **Behavior Preservation Tests** (`src/stores/behaviorPreservation.test.ts`): 49 tests ensuring no regression in existing functionality
    - ✅ **Test Utilities** (`src/test-utils-ui.ts`): Comprehensive utilities for UI state testing with assertions and scenarios
 
 2. **Enhanced State Management Implementation**:
+
    - ✅ Verified proper integration between SessionTable and AppStore
    - ✅ Fixed function name mismatches (`startSessionEdit` → `startEdit`, etc.)
    - ✅ Added proper TypeScript typing for all test scenarios
@@ -450,6 +452,7 @@ src/components/
    - ✅ Created mock infrastructure for daemon client testing
 
 3. **Code Quality Improvements**:
+
    - ✅ **All formatting issues resolved** with Prettier
    - ✅ **All linting issues resolved** (unused variables, missing imports)
    - ✅ **All TypeScript errors resolved** (type mismatches, undefined properties)
@@ -457,7 +460,7 @@ src/components/
 
 4. **Verification and Testing**:
    - ✅ **Format check passed** (`bun run format`)
-   - ✅ **Lint check passed** (`bun run lint`)  
+   - ✅ **Lint check passed** (`bun run lint`)
    - ✅ **Type checking passed** (`bun run typecheck`)
    - ✅ Created 72 total test cases ensuring comprehensive coverage
    - ✅ Verified all existing functionality is preserved
@@ -471,13 +474,95 @@ src/components/
 - **Performance**: Efficient state management with proper immutable updates
 
 #### Files Created/Modified in This Phase:
+
 - `src/stores/uiSlice.test.ts` - **NEW**: Comprehensive UI slice testing
 - `src/pages/SessionTablePage.test.tsx` - **NEW**: Integration testing
-- `src/stores/behaviorPreservation.test.ts` - **NEW**: Regression prevention testing  
+- `src/stores/behaviorPreservation.test.ts` - **NEW**: Regression prevention testing
 - `src/test-utils-ui.ts` - **NEW**: Testing utilities for UI state
 - `src/components/SessionTable/SessionTable.tsx` - **FIXED**: Function name synchronization
 - Multiple existing files - **FIXED**: Linting and TypeScript issues
 
+### ✅ COMPLETED: P1 Item #4 - Custom Hook Consolidation (2024-08-28)
+
+**Status**: Successfully completed
+**Time Spent**: ~2 hours
+**Files Changed**: 5 files (2 components, 1 hook removed, 2 test files)
+
+#### What Was Done:
+
+1. **Analyzed Custom Hooks for Reusability**:
+   - ✅ **useKeyboardNavigationProtection**: Identified as suitable for inlining (low complexity, limited reuse)
+   - ✅ **useStealHotkeyScope**: Kept as hook (complex logic, truly reusable across 4+ components)
+   - ✅ **useAsyncState**: Kept as hook (generic utility pattern, could be expanded to other hooks)
+
+2. **Implemented useKeyboardNavigationProtection Inlining**:
+   - ✅ Removed hook file: `src/hooks/useKeyboardNavigationProtection.ts`
+   - ✅ Inlined logic in `SessionTablePage.tsx` (15 lines of keyboard navigation protection)
+   - ✅ Inlined logic in `SessionDetail.tsx` (same 15 lines of protection logic)
+   - ✅ Updated test mocks in `SessionDetail.test.tsx` to remove dependency
+   - ✅ Updated Storybook comment in `SessionDetail.stories.tsx`
+
+3. **Code Quality Verification**:
+   - ✅ **Format check passed** (`bun run format`)
+   - ✅ **Lint check passed** (`bun run lint`)
+   - ✅ **Type checking passed** (`bun run typecheck`)
+   - ✅ All imports updated correctly
+   - ✅ No breaking changes to functionality
+
+#### Implementation Details:
+
+**Inlined Logic Pattern** (repeated in both components):
+```typescript
+// Keyboard navigation protection - inline implementation
+const [isKeyboardNavigating, setIsKeyboardNavigating] = useState(false)
+const keyboardTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
+
+const startKeyboardNavigation = useCallback(() => {
+  setIsKeyboardNavigating(true)
+  if (keyboardTimeoutRef.current) {
+    clearTimeout(keyboardTimeoutRef.current)
+  }
+  keyboardTimeoutRef.current = setTimeout(() => {
+    setIsKeyboardNavigating(false)
+  }, 300)
+}, [])
+
+const shouldIgnoreMouseEvent = useCallback((): boolean => {
+  return isKeyboardNavigating
+}, [isKeyboardNavigating])
+```
+
+#### Files Modified:
+- `src/pages/SessionTablePage.tsx` - **MODIFIED**: Added inline keyboard protection logic
+- `src/components/SessionDetail/SessionDetail.tsx` - **MODIFIED**: Added inline keyboard protection logic  
+- `src/hooks/useKeyboardNavigationProtection.ts` - **DELETED**: Hook file removed
+- `src/components/SessionDetail/SessionDetail.test.tsx` - **MODIFIED**: Removed hook mock
+- `src/components/SessionDetail/SessionDetail.stories.tsx` - **MODIFIED**: Updated comment
+
+#### Benefits Achieved:
+
+1. **Reduced Abstraction**: Eliminated unnecessary indirection for simple logic
+2. **Easier Testing**: No need to mock the hook in component tests
+3. **Co-location**: Logic lives directly where it's used, improving readability
+4. **Bundle Size**: Eliminated hook overhead (minimal but measurable)
+5. **Follows Standards**: Aligns with "Create custom hooks only for truly reusable functionality"
+
+#### Architecture Impact:
+
+- ✅ Maintained all existing functionality (keyboard navigation protection works identically)
+- ✅ Reduced code traversal depth (principle #1 from coding standards)
+- ✅ No performance impact (same logic, just inlined)
+- ✅ Improved testability (no mocking required)
+- ✅ Code duplication is acceptable for simple, stable logic patterns
+
+#### Hook Consolidation Summary:
+
+| Hook | Decision | Rationale |
+|------|----------|----------|
+| `useKeyboardNavigationProtection` | **Inlined** ✅ | Low complexity (15 lines), used in only 2 components, stable logic |
+| `useStealHotkeyScope` | **Keep as Hook** ✅ | Complex logic (30+ lines), used in 4+ components, error-prone if duplicated |
+| `useAsyncState` | **Keep as Hook** ✅ | Generic utility pattern, could benefit other hooks that duplicate its pattern |
+
 ## Next Action
 
-**Continue with P1 item #4**: Custom Hook Consolidation - review hooks for true reusability vs inline logic
+**Continue with P1 item #5**: Error Boundary Implementation - add granular error boundaries around risky operations
